+++ app/src/main/java/com/triadss/doctrack2/activity/healthprof/fragments/appointments/HealthProfessionalPending.java
@@ -1,4 +1,4 @@
-package com.triadss.doctrack2.activity.healthprof;
+package com.triadss.doctrack2.activity.healthprof.fragments.appointments;
 
 import static androidx.constraintlayout.helper.widget.MotionEffect.TAG;
