+++ app/src/main/java/com/triadss/doctrack2/activity/healthprof/fragments/appointments/HealthProfessionalUpcoming.java
@@ -1,4 +1,4 @@
-package com.triadss.doctrack2.activity.healthprof;
+package com.triadss.doctrack2.activity.healthprof.fragments.appointments;
 
 import android.content.Context;
 import android.content.SharedPreferences;
